feat: creating componentization at "navigation"
