+++ src/components/DrawerContent/DrawerContent.tsx
@@ -1,26 +1,31 @@
-import { DrawerContentScrollView, DrawerItem } from "@react-navigation/drawer";
 import { View, Image } from "react-native";
+import { useNavigation } from "@react-navigation/native";
+import { DrawerContentScrollView, DrawerItem } from "@react-navigation/drawer";
 
 import { styles } from "./styles";
-import { buttons } from "./buttons";
+import { BUTTONS } from "./buttons";
 
+import { NavigationProps } from "../../navigation/types";
 import iconGerandoFalcoes from "../../../assets/images/iconGerandoFalcoes.png";
 
-export const DrawerContent: React.FC = (props: T) => {
+export const DrawerContent: React.FC = () => {
+
+    const navigation =  useNavigation<NavigationProps>()
+
     return (
-        <DrawerContentScrollView {...props} >
+        <DrawerContentScrollView>
             <View
                 style={styles.container}
             >
                 <View>
                     <Image source={iconGerandoFalcoes} style={styles.iconGF} />
                 </View>
-                {buttons.map((button, index) => (
+                {BUTTONS.map((button, index) => (
                     <DrawerItem
                         key={index}
                         label={button.label}
                         icon={() => <Image source={button.icon} style={styles.iconDrawerItem} />}
-                        onPress={() => props.navigation.navigate(button.navigateTo)}
+                        onPress={() => navigation.navigate(button.navigateTo)}
                         style={styles.drawerItem}
                         labelStyle={styles.drawerItemLabel}
                     />
