+++ src/navigation/Drawer/index.ts
@@ -1,0 +1,1 @@
+export * from './DrawerMenu';
