+++ src/navigation/index.ts
@@ -1,1 +1,1 @@
-export * from './Navigation'
+export * from './Navigation';
